requirements.txt
dashscope==1.14.1
faster_whisper==0.10.0
Flask==2.2.2
ImageHash==4.3.1
jieba==0.42.1
llama_index==0.12.19
modelscope==1.11.0
numpy==1.23.5
openai==1.63.2
opencv_contrib_python==4.9.0.80
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
paddlex==3.0.0b1
Pillow==11.1.0
pywebview==5.2
scikit_learn==1.3.0
scipy==1.15.2
tqdm==4.66.2
